Refactor: Changed project name as Example for true presentation
