+++ src/main/java/com/enesthedev/example/Application.java
@@ -1,4 +1,4 @@
-package com.enesthedev.customer;
+package com.enesthedev.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
